Update experimentalURLGrabFile.c
Added if statement to check in the HTML parser if the target link is in there; if so, returns an exit success.  

Additionally added functionality in main to take the target article, eliminate the prefix to just get the title from the article link, then adds that as an input into the HTML Parser function for which it can then check for its presence.