Update experimentalURLGrabFile.c
Functionalized the URL grabber so that it can work modularly.  Haven't done much testing with it yet but it should be all happy and good.  Should work just fine.  

I think I will soon create a file pool assignment function so that the files can be distributed to individual threads of execution.